requirements.txt
discord.py==2.3.2
aiohttp
Pillow==10.1.0
python-dotenv==1.0.0
replicate==0.22.0
asyncio==3.4.3
anthropic==0.18.0
openai==1.6.1
elevenlabs==0.2.27
python-dateutil==2.8.2
requests
typing-extensions>=4.5.0  # Required for better type hints
psutil>=5.9.0  # Required for process management
playsound
gradio_client  # Required for Zonos TTS
pydub
opencv-python
